fix: address PR review — rem_euclid for negative timestamps, consolidate date extraction, fix combinator JSON format
- Use div_euclid/rem_euclid in extract_date_from_column for correct
  handling of pre-1970 nanosecond timestamps
- Add String type support to extract_date_from_column and remove
  duplicate extract_naive_date from signals/mod.rs
- Update combinator descriptions to show actual JSON wire format
  instead of Rust struct syntax
