+++ src/signals/mod.rs
@@ -14,6 +14,7 @@ use anyhow::Result;
 use chrono::NaiveDate;
 use polars::prelude::*;
 
+use crate::engine::event_sim::extract_date_from_column;
 use helpers::SignalFn;
 use registry::{build_signal, SignalSpec};
 
@@ -36,71 +37,12 @@ pub fn active_dates(
 
     for i in 0..ohlcv_df.height() {
         if bool_ca.get(i) == Some(true) {
-            let date = extract_naive_date(col, i)?;
+            let date = extract_date_from_column(col, i)?;
             result.insert(date);
         }
     }
 
     Ok(result)
-}
-
-/// Extract a `NaiveDate` from a column value at a given index.
-/// Handles Date, Datetime, and String column types.
-fn extract_naive_date(col: &Column, idx: usize) -> Result<NaiveDate> {
-    match col.dtype() {
-        DataType::Date => {
-            let days = col.date()?.phys.get(idx);
-            match days {
-                Some(d) => {
-                    let date = NaiveDate::from_num_days_from_ce_opt(d + 719_163)
-                        .ok_or_else(|| anyhow::anyhow!("Invalid date at index {idx}"))?;
-                    Ok(date)
-                }
-                None => anyhow::bail!("Null date at index {idx}"),
-            }
-        }
-        DataType::Datetime(tu, _) => {
-            let val = col.datetime()?.phys.get(idx);
-            match val {
-                Some(v) => {
-                    let ndt = match tu {
-                        TimeUnit::Milliseconds => {
-                            chrono::DateTime::from_timestamp_millis(v).map(|dt| dt.naive_utc())
-                        }
-                        TimeUnit::Microseconds => {
-                            chrono::DateTime::from_timestamp_micros(v).map(|dt| dt.naive_utc())
-                        }
-                        TimeUnit::Nanoseconds => {
-                            let secs = v / 1_000_000_000;
-                            let nsecs = (v % 1_000_000_000) as u32;
-                            chrono::DateTime::from_timestamp(secs, nsecs).map(|dt| dt.naive_utc())
-                        }
-                    };
-                    match ndt {
-                        Some(dt) => Ok(dt.date()),
-                        None => anyhow::bail!("Invalid datetime value at index {idx}"),
-                    }
-                }
-                None => anyhow::bail!("Null datetime at index {idx}"),
-            }
-        }
-        DataType::String => {
-            let str_val = col.str()?.get(idx);
-            match str_val {
-                Some(s) => {
-                    // Try YYYY-MM-DD first, then YYYY-MM-DDTHH:MM:SS
-                    NaiveDate::parse_from_str(s, "%Y-%m-%d")
-                        .or_else(|_| {
-                            chrono::NaiveDateTime::parse_from_str(s, "%Y-%m-%dT%H:%M:%S")
-                                .map(|dt| dt.date())
-                        })
-                        .map_err(|e| anyhow::anyhow!("Cannot parse date '{s}': {e}"))
-                }
-                None => anyhow::bail!("Null string date at index {idx}"),
-            }
-        }
-        other => anyhow::bail!("Unsupported column type for date extraction: {other:?}"),
-    }
 }
 
 #[cfg(test)]
